models: Split models into smaller modules
